created reusable RotatingImmgaes and RandomImageGenerator components
